reporting/sql-reports/requirements.txt
click
dnspython3
mysql-connector-python
openpyxl
pandas
python-dotenv
pytz
xlsxwriter
